chore: setup admin workspace
