+++ .github/actions/setup-rust/action.yml
@@ -1,0 +1,32 @@
+name: Setup Rust
+description: Setup Rust with Cargo cache and optional cargo-lambda
+
+inputs:
+  cargo-lambda:
+    description: "Whether to install cargo-lambda"
+    required: false
+    default: "false"
+
+runs:
+  using: composite
+  steps:
+    - name: Cache Cargo
+      uses: actions/cache@27d5ce7f107fe9357f9df03efb73ab90386fccae # v5.0.5
+      with:
+        path: |
+          ~/.cargo/registry
+        key: ${{ runner.os }}-cargo-${{ hashFiles('**/Cargo.lock') }}
+        restore-keys: |
+          ${{ runner.os }}-cargo-
+
+    - name: Cache pip
+      if: inputs.cargo-lambda == 'true'
+      uses: actions/cache@27d5ce7f107fe9357f9df03efb73ab90386fccae # v5.0.5
+      with:
+        path: ~/.cache/pip
+        key: ${{ runner.os }}-pip-cargo-lambda
+
+    - name: Install cargo-lambda
+      if: inputs.cargo-lambda == 'true'
+      shell: bash
+      run: pip3 install cargo-lambda
