+++ apps/admin/src/main.rs
@@ -1,0 +1,10 @@
+use lambda_runtime::{run, service_fn, tracing, Error};
+mod handler;
+use handler::function_handler;
+
+#[tokio::main]
+async fn main() -> Result<(), Error> {
+    tracing::init_default_subscriber();
+
+    run(service_fn(function_handler)).await
+}
