+++ packages/auth/package.json
@@ -19,9 +19,9 @@
     }
   },
   "dependencies": {
-    "@better-auth/drizzle-adapter": "https://pkg.pr.new/better-auth/better-auth/@better-auth/drizzle-adapter@6913",
+    "@better-auth/drizzle-adapter": "https://pkg.pr.new/better-auth/better-auth/@better-auth/drizzle-adapter@9489",
     "@tokistack/db": "workspace:*",
-    "better-auth": "1.5.0-beta.13"
+    "better-auth": "https://pkg.pr.new/better-auth/better-auth/better-auth@9489"
   },
   "devDependencies": {
     "@tokistack/typescript-config": "workspace:*",
